feat(gs): make sure old charts get v1 added
