+++ app/src/lib/getDefaultChart.ts
@@ -1,0 +1,5 @@
+import { getDefaultText } from "./getDefaultText";
+
+export function getDefaultChart() {
+  return `${getDefaultText()}\n\n=====\n{}\n=====`;
+}
