+++ app/src/lib/getDefaultText.ts
@@ -1,0 +1,20 @@
+import { t } from "@lingui/macro";
+
+export function getDefaultText() {
+  const defaultText = `${t`This app works by typing`}
+  ${t`Indenting creates a link to the current line`}
+  ${t`any text: before a colon creates a label`}
+  ${t`Create a link directly using the exact label text`}
+    ${t`like this: (This app works by typing)`}
+    ${t`[custom ID] or`}
+      ${t`by adding an %5BID%5D and referencing that`}
+        ${t`like this: (custom ID) // You can also use single-line comments`}
+/*
+${t`or`}
+${t`multiline`}
+${t`comments`}
+
+${t`Have fun! 🎉`}
+*/`;
+  return defaultText;
+}
