+++ app/src/lib/prepareChart/examples/example3
