+++ app/src/lib/prepareChart/examples/example8
@@ -1,0 +1,7 @@
+old
+  [x] parser
+=====
+{
+  "parser": "v1"
+}
+=====
