+++ app/src/lib/prepareChart/examples/example9
@@ -1,0 +1,3 @@
+This app works by typing
+  Indenting creates a link to the current line
+  any text: before a colon creates a label
